+++ internal/renderers/skia/Cargo.toml
@@ -1,0 +1,71 @@
+# Copyright © SixtyFPS GmbH <info@slint-ui.com>
+# SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-Slint-commercial
+
+[package]
+name = "i-slint-renderer-skia"
+version = "0.3.3"
+authors = ["Slint Developers <info@slint-ui.com>"]
+edition = "2021"
+license = "GPL-3.0-only OR LicenseRef-Slint-commercial"
+description = "Skia based renderer for Slint"
+repository = "https://github.com/slint-ui/slint"
+homepage = "https://slint-ui.com"
+build = "build.rs"
+rust-version.workspace = true
+
+[lib]
+path = "lib.rs"
+
+# Note, these features need to be kept in sync (along with their defaults) in
+# the C++ crate's CMakeLists.txt
+[features]
+wayland = ["glutin/wayland"]
+x11 = ["glutin/x11"]
+opengl = ["skia-safe/gl", "glow"]
+default = []
+
+[dependencies]
+i-slint-core = { version = "=0.3.3", path = "../../../internal/core", features = [ "box-shadow-cache" ] }
+i-slint-core-macros = { version = "=0.3.3", path = "../../../internal/core-macros" }
+i-slint-common = { version = "=0.3.3", path = "../../../internal/common" }
+
+const-field-offset = { version = "0.1", path = "../../../helper_crates/const-field-offset" }
+vtable = { version = "0.1.6", path = "../../../helper_crates/vtable" }
+
+cfg-if = "1"
+derive_more = "0.99.5"
+lyon_path = "1.0"
+once_cell = "1.5"
+pin-weak = "1"
+scoped-tls-hkt = "0.1"
+dark-light = "0.2.2"
+instant = "0.1"
+raw-window-handle = { version = "0.5", features = ["alloc"] }
+
+skia-safe = { version = "0.56.1", features = ["textlayout"] }
+glow = { version = "0.11", optional = true }
+unicode-segmentation = { version = "1.8.0" }
+
+glutin = { version = "0.30", default-features = false, features = ["egl", "wgl"] }
+
+[target.'cfg(target_family = "windows")'.dependencies]
+winapi = { version = "0.3", features = ["dwrite", "d3d12", "dxgi", "dxgi1_2", "dxgi1_3", "dxgi1_4", "d3d12sdklayers", "synchapi"] }
+skia-safe = { version = "0.56.1", features = ["d3d"] }
+wio = { version = "0.2.2" }
+
+[target.'cfg(target_os = "macos")'.dependencies]
+cocoa = { version = "0.24.0" }
+core-foundation = { version = "0.9.1" }
+core-text = { version = "19.1.0" }
+metal = { version = "0.24.0" }
+# Use the same version of foreign-types as the metal crate uses.
+foreign-types = { version = "0.3.2" }
+objc = { version = "0.2.7" }
+core-graphics-types = { version = "0.1.1" }
+skia-safe = { version = "0.56.1", features = ["metal"] }
+
+[target.'cfg(not(any(target_os = "macos", target_family = "windows")))'.dependencies]
+skia-safe = { version = "0.56.1", features = ["gl"] }
+
+[build-dependencies]
+cfg_aliases = "0.1.0"
