+++ internal/renderers/skia/cached_image.rs
